requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Network Scanning
python-nmap==0.7.1
masscan==0.1.6
scapy==2.5.0
requests==2.31.0
aiohttp==3.9.1

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.8
groq==0.4.1

# Active Directory Integration
ldap3==2.9.1
msal==1.25.0

# External API Integrations
httpx==0.25.2
tenacity==8.2.3

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# SSL/TLS
pyopenssl==23.3.0
certbot==2.7.0

# MFA Support
pyotp==2.9.0
qrcode[pil]==7.4.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
